requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Security-Too==5.3.1
Flask-Bootstrap==3.3.7.1
python-dotenv==1.0.0